requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
idna==3.3
Pillow==8.4.0
python-dateutil==2.8.2
python-slugify==5.0.2
pyusb==1.2.1
requests==2.26.0
six==1.16.0
text-unidecode==1.3
tqdm==4.62.3
urllib3==1.26.7
